test(auto-reply): split thinking defaults coverage
